Create CountDigits.java
This is one of the very fundamental question asked in the coding Interviews , To write a program to count the no. of digits of the number (Say n) which we are taking as an input from the user. 
Here i come up with the super optimised solution for that problem .
Algorithm:- 
1. Check if digit exists (n>0).
2. Remove the last digit (n=n/10)
3. Increase the value of Count by 1.(count++)
4. Repeat Step 1 To Step 3 till digit's Remains (loop)
Example Input :- 1989 ,Output:-4.